pyproject.toml
[project]
name = "memobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.13.0",
    "annotated-types==0.7.0",
    "av>=10,<13",
    "backoff==2.2.1",
    "celery==5.3.4",
    "deepface>=0.0.79",
    "diskcache==5.6.3",
    "fastapi==0.104.1",
    "moviepy>=1.0.3",
    "noisereduce>=3.0.0",
    "gdown>=5",
    "google-cloud-aiplatform>=1.115.0",
    "google-genai>=1.2.0",
    "graphiti-core",
    "httpx==0.25.2",
    "neo4j>=5.14.0,<7",
    "numpy>=1.26.2,<2",
    "openai>=1.3.7,<3",
    "opencv-python>=4.8,<5",
    "orjson>=3.11.5,<4",
    "packaging>=24.2,<26",
    "pandas>=2,<3",
    "passlib[bcrypt]==1.7.4",
    "pgvector==0.2.4",
    "pinecone>=7,<9",
    "pinecone-plugin-assistant==3.0.1",
    "pinecone-plugin-interface>=0.0.7,<1",
    "posthog==7.7.0",
    "prometheus-client==0.19.0",
    "psycopg2-binary==2.9.9",
    "pyannote-audio==3.3.2",
    "pyarrow==17.0.0",
    "pyaudio>=0.2.13",
    "pydantic==2.12.5",
    "pydantic-settings==2.1.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "python-dateutil>=2.8.2,<3",
    "python-dotenv>=1.0.1,<2",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "python-speech-features>=0.6",
    "redis==5.0.1",
    "requests>=2.32.3,<3",
    "scenedetect>=0.6",
    "sentence-transformers==2.2.2",
    "silero-vad>=6.2.0",
    "six==1.17.0",
    "sounddevice==0.5.5",
    "sqlalchemy==2.0.23",
    "tenacity==9.1.2",
    "tensorflow==2.19.0",
    "tensorflow-metal>=1.2.0 ; sys_platform == 'darwin'",
    "tf-keras>=2.19.0",
    "torch>=2.10.0",
    "tqdm>=4.67.1,<5",
    "typing-extensions==4.15.0",
    "uvicorn[standard]==0.24.0",
    "websockets>=12,<17",
    "pvporcupine>=4.0.2",
]

[tool.uv.sources]
graphiti-core = { git = "https://github.com/Cathy-ZYX119/graphiti.git", rev = "cedc7a08600d5573cd2b2b7fcaa796f5876240ba" }
